% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nullnetr.R
\docType{package}
\name{nullnetr}
\alias{nullnetr}
\alias{nullnetr-package}
\title{nullnetr: tools for using null models to analyse the structure of ecological
  networks and identify resource selection}
\description{
nullnetr provides a set of functions for running null models, analysing and
  plotting the results, and exporting the outputs for use with other network
  packages in R.
}
\references{
Vaughan, I.P., Gotelli, N.J., Memmott, J., Pearson, C.E.,
  Woodward, G. and Symondson, W.O.C. (2017) nullnetr: an R package using null
  models to analyse the structure of ecological networks and identify
  resource selection (submitted)
}

