% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_preferences.R
\name{plot_preferences}
\alias{plot_preferences}
\title{Plot the resource preferences of a consumer}
\usage{
plot_preferences(nullnet, node, signif.level = 0.95, bar.type = "counts",
  res.order = NULL, bar.colours = NULL, bar.colors = NULL, ...)
}
\arguments{
\item{nullnet}{An object of class nullnet from \code{generate_null_net}}

\item{node}{A string specifying the consumer node (species) whose
preferences will be plotted}

\item{signif.level}{An optional value specifying the threshold used for
testing for 'significant' deviations from the null model.  Defaults to 0.95}

\item{bar.type}{Optional string to specify what the bars show. The default
(\code{bar.type = "counts"}) is the total number of interactions or mean
interaction strength, depending upon which one was requested in the
original call to \code{\link{generate_null_net}} using the
\code{summary.type} argument.  The alternative is the standardised effect
size (\code{bar.type = "SES"}).  With \code{bar.type = "counts"}, confidence
limits will be shown as error bars, whilst for the standardised effect size,
dashed lines are added at +2 and -2, which is approximately equivalent to 5\%
significance level.  Both types of bars will be coloured if \code{bar.colours}
is specified.}

\item{res.order}{An optional data frame used to set the order in which the
resource species are plotted.  Should have two columns: the first listing
the resource species (names must be identical to those used by
\code{generate_null_net}) and the second the plotting order (bars will be
plotted in ascending order)}

\item{bar.colours}{An optional character vector of length three specifying
the colours with which to fill bars representing interactions that are
weaker than expected, consistent with the null model and stronger than
expected (in that order).  If this argument is not included, the bars will
be unfilled.}

\item{bar.colors}{Alternative spelling for \code{bar.colours}}

\item{...}{Other arguments to control basic plotting functions in R, such as
\code{xlab}, \code{ylab}, \code{main}, \code{xlim} and \code{ylim}.}
}
\description{
Takes a 'nullnet' object from running a null model with
  \code{generate_null_net} and plots the observed and expected link
  strengths for every resource for a selected consumer species (node in
  the network).  The function is based on \code{\link[graphics]{barplot}}, so
  see \code{?barplot} for further options for customising the plots.
}
\details{
Plots the preferences for individual consumer species, following the
  basic format of King \emph{et al}. (2010; Figure 3) and Davey \emph{et al}.
  (2013; Figure 3).
}
\examples{
null.1 <- generate_null_net(WelshStreams[, 2:18], WelshStreams.prey[, 2:17],
                            sims = 10, c.samples = WelshStreams[, 1],
                            r.samples = WelshStreams.prey[, 1])
# Generate plot using three-level diverging colour scheme to fill the bars
#   (Brewer 2017), increasing the lower margin so that names fit
par(mar = c(9, 4, 4, 2) + 0.1)
plot_preferences(null.1, "Dinocras", signif.level = 0.95,
                 bar.colours = c("#998EC3", "#F7F7F7", "#F1A340"),
                 bar.type = "counts", ylab = "Num. of visits")

# Same results, this time showing the standardised effect sizes
plot_preferences(null.1, "Rhyacophila", signif.level = 0.95,
                 bar.colours = c("#998EC3", "#F7F7F7", "#F1A340"),
                 bar.type = "SES", ylab = "SES")

}
\references{
Brewer, C.A. (2017) \url{http://www.ColorBrewer.org}

  Davey, J.S., Vaughan, I.P., King, R.A., Bell, J.R., Bohan, D.A., Bruford,
  M.W., Holland, J.M. & Symondson, W.O.C. (2013) Intraguild predation in
  winter wheat: prey choice by a common epigeal carabid consuming spiders.
  \emph{Journal of Applied Ecology}, \strong{50}, 271--279.

  King, R.A, Vaughan, I.P., Bell, J.R., Bohan, D.A, & Symondson, W.O.C.
  (2010) Prey choice by carabid beetles feeding on an earthworm community
  analysed using species- and lineage-specific PCR primers. \emph{Molecular
  Ecology}, \strong{19}, 1721--1732.

  Vaughan, I.P., Gotelli, N.J., Memmott, J., Pearson, C.E.,
  Woodward, G. and Symondson, W.O.C. (2017) nullnetr: an R package using null
  models to analyse the structure of ecological networks and identify
  resource selection (submitted)
}
\seealso{
\code{\link{generate_null_net}}, \code{\link{plot_bipartite}}
}

